Bak for key report generators
SLMauto debug v2 is the final notebook version for generating ASTC reports, generating from a testplan all the key info, pasting into excel, and using excel to output PDFs. 
ASTC calc from testplan is a prelim version I've been developing to do the ASTC calc in python itself. That is still needing dev, as we still need to connect it to the reporting and subsequent PDF generation of a plot and reporting info.